Refactor [#9] 인터페이스 분리로 인한 네이밍 변경
