--- morib/src/main/java/org/morib/server/api/homeViewApi/service/create/CreateTaskService.java
@@ -1,5 +1,0 @@
-package org.morib.server.api.homeViewApi.service.create;
-
-public interface CreateTaskService {
-    void execute();
-}
