+++ morib/src/main/java/org/morib/server/api/homeViewApi/service/create/task/CreateTaskService.java
@@ -1,0 +1,5 @@
+package org.morib.server.api.homeViewApi.service.create.task;
+
+public interface CreateTaskService {
+    void create();
+}
