+++ morib/src/main/java/org/morib/server/api/homeViewApi/service/create/todo/CreateTodoService.java
@@ -1,0 +1,5 @@
+package org.morib.server.api.homeViewApi.service.create.todo;
+
+public interface CreateTodoService {
+    void create();
+}
