--- morib/src/main/java/org/morib/server/api/homeViewApi/service/toggle/ToggleTaskStatusServiceImpl.java
@@ -1,20 +1,0 @@
-package org.morib.server.api.homeViewApi.service.toggle;
-
-import lombok.RequiredArgsConstructor;
-import org.morib.server.domain.task.TaskOperator;
-import org.morib.server.domain.task.infra.TaskGateway;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class ToggleTaskStatusServiceImpl implements ToggleTaskStatusService{
-    private final TaskGateway taskGateway;
-    private final TaskOperator taskOperator;
-
-    @Override
-    public void execute() {
-        // gateway로 task 꺼내와서
-        // switch로 상태 변경 후
-        // gateway로 변경된 task 저장
-    }
-}
